+++ scripts/install-services.sh
@@ -54,6 +54,7 @@ modules=(
     server/templates/messages/batch_completed.html
     server/templates/messages/attempt_started.html
     server/templates/messages/reply_received.html
+    server/templates/messages/prompt_sent.html
     server/templates/messages/tool_completed.html
     server/templates/messages/submission_accepted.html
     server/templates/messages/tool_received.html